etl/requirements.txt
requests
pandas
google-cloud-bigquery
python-dotenv 
